Version 1.0.0h - see change log v1.0.0

+++ src/classes/Tool/CLI/Table.php
@@ -21,7 +21,7 @@
  *	@package		CeusMedia.Hymn.Tool.CLI
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@copyright		2014-2024 Christian Würker
- *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
+ *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  */
 /**
@@ -31,7 +31,7 @@
  *	@package		CeusMedia.Hymn.Tool.CLI
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@copyright		2014-2024 Christian Würker
- *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
+ *	@license		https://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Hymn
  *	@todo			code documentation
  */
@@ -169,7 +169,7 @@ class Hymn_Tool_CLI_Table
 
 	protected function line( $sign = '-', int $lineLength = 76 ): string
 	{
-		$steps	= floor( $lineLength / $this->strlen( $sign ) );
+		$steps	= (int) floor( $lineLength / $this->strlen( $sign ) );
 		return str_repeat( $sign, $steps );
 	}
